feat: implement pricing page and subscription cards (#11)
